refactor(desktop): split app.go into per-domain files (<500 lines each)
app.go was 1,113 lines. Move the Wails binding methods into
same-package files by domain — no behavior change, just file layout:

- app.go (257): struct, lifecycle, credentials, api()/enabled() helpers
- app_analyzer.go (201): action plan, analyzer/deterministic rules, queries
- app_features.go (219): threading, obsidian/slack, labels, links, drafts
- app_mail.go (460): keymap, accounts, mailbox core, AI

go build + vet + gofmt clean.
